requirements.txt
tensorflow==2.16.1  
opencv-python==4.10.0.84  
Pillow==11.3.0  
pyzbar==0.1.9  
streamlit==1.50.0  
tensorboard==2.16.1
numpy==1.26.4
requests==2.32.5
onnx==1.17.0
tf2onnx==1.16.1
gradio_client

# For Raspberry Pi, use opencv-python-headless instead:
# opencv-python-headless==4.8.1.78

# For Jetson Nano, TensorFlow should be installed separately:
# See: https://developer.nvidia.com/embedded/downloads

# "Installing TensorFlow for Jetson..."
# "Please install manually from: https://developer.nvidia.com/embedded/downloads"
# "Or use: pip install --extra-index-url https://developer.download.nvidia.com/compute/redist/jp/v461 tensorflow"